added sending contract to the buyer functionalities

+++ app/Http/Controllers/BuyEstateController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Contract;
 use App\Models\Estate;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -19,10 +20,29 @@ class BuyEstateController extends Controller
 
     public function cancel(Estate $estate) {
 
-        $user = User::find(auth()->id());
+        $contracts = Contract::all()->pluck('estate_id');
         
-        $user->boughtEstates()->detach($estate);
+        if($contracts->contains($estate->id)) {
+            echo "Contract is written";
+        } else {
+            $user = User::find(auth()->id());
+        
+            $user->boughtEstates()->detach($estate);
+    
+            return redirect()->route('index')->with('success', 'Your request was canceled!');
+        }
+    }
 
-        return redirect()->route('index')->with('success', 'Your request was canceled!');
+    public function sent_requests(User $user) {
+        $this->authorize('update', $user);
+
+        $bought = $user->boughtEstates()->pluck('estate_id');
+
+        $estates = Estate::whereIn('id', $bought);
+
+        return view('index', [
+            'estates' => $estates->paginate(5),
+            'user' => $user
+        ]);
     }
 }
